Quite happy with this commit, it brings some nice improvements.
* Fixes many marhsaling bugs:

  * Godot likes to expose on both ends types
    like uint32, but expects the marshaling to take place as int64
    values - fun for the whole family.

  * Swift helpfully turns a 'enum Foo: Int' into the most compact
    representation (cool in structures), not so cool when it happens
    on the stack, and you assume that you can pass a pointer to the
    enum and get 8 solid bytes: not so, this uses one byte if it does
    not need more space

* Added a "ClassServices" type to start the process of surfacing
  methods and signals from a Swift type.  The bundled example shows
  how to define Signals and Methods, and how to invoke them.

+++ Generator/Generator/TypeHelpers.swift
@@ -12,7 +12,7 @@ import Foundation
 func BuiltinJsonTypeToSwift (_ type: String) -> String {
     switch type {
     case "float": return "Float"
-    case "int": return "Int64"
+    case "int": return "Int"
     case "bool": return "Bool"
     default:
         return type
